requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
PyPDF2
streamlit
transformers
torch